+++ src/result.rs
@@ -1,3 +1,3 @@
 use crate::error::Error;
 
-type Result<T> = std::result::Result<T, Error>;
+pub type Result<T> = std::result::Result<T, Error>;
