+++ src/types/traits.rs
@@ -1,0 +1,1 @@
+
